Add initial test application
Test application can use simDetectorConfigure function, NDStdArrays
plugin and also the plugin being developed itself.

I couldn't think of a way to add only modules paths such as BUSY and
ADSIMDETECTOR only to testApp, but that shouldn't be a problem since
the libs and dbds are only added in testApp/src/Makefile.

The st.cmd has two stdArrays: one for connecting to the simdetector
and seeing the raw image and one for connecting to the plugin and seeing
the image that comes out of it, if any iimage comes out at all.

Remember to keep hydrated.